do the thing with core/edm
