+++ src/core/edm/reducers/index.js
@@ -1,0 +1,9 @@
+/*
+ * @flow
+ */
+
+import reducer from './reducer';
+
+export { default as getEntityDataModelTypesReducer } from './getEntityDataModelTypesReducer';
+
+export default reducer;
