+++ src/core/edm/sagas/index.js
@@ -1,0 +1,5 @@
+/*
+ * @flow
+ */
+
+export * from './getEntityDataModelTypes';
